requirements.txt
Pyrogram==1.2.9
python-dotenv==0.19.1
